+++ wp-content/themes/bplactive/inc/resources-vault/templates/resource-modal.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * Resource modal template (included in modal.php)
+ */
+if ( ! defined( 'ABSPATH' ) ) exit;
+// Modal markup is output by modal.php